// src/Customer/Sell/types/product.type.ts
export type TFormProduct = {
    product_name: string

    product_price: number | null
    product_thumb_image: { secure_url: string; public_id: string }
    product_image_desc: string[]
    product_id: string
}

export type TFormBook = {
    publishing: string
    author: string
    page_number: number
    description: string
}

//@type upload 1 hình
export type UploadImage = {
    product_id: string
    product_thumb_image: { secure_url: string; public_id: string }
    FileName: string
    FileLength: number
}

//@type upload 1 mảng
export type UploadImages = {
    secure_url: string
    public_id: string
}